Add data request for multi landlord.
